🚚 Rename: activity 도메인과의 연관성을 고려해 파일 위치 변경
